% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigPCA.R
\name{bigPCA}
\alias{bigPCA}
\title{Run a PCA over a set of eBird records}
\usage{
bigPCA(comparisons, read.wd, write.wd, columns, scale.center = FALSE, axes)
}
\arguments{
\item{comparisons}{List where each element is a named character vector. The name of
the element specifies what the PCA results will be called, and the character vector
lists the species that will be combined before the PCA is run.}

\item{read.wd}{Working directory where the eBird records with environmental data are
stored.}

\item{write.wd}{Working directory where PCA results will be saved.}

\item{columns}{Character vector of column names in the existing eBird records that
you will run the PCA over.}

\item{scale.center}{Default is FALSE, i.e. a covariance matrix PCA will be run. To
run a correlation matrix PCA, set to TRUE. Scaling and centering are accomplished
manually before the PCA is run, which I think might be slow to run for really large
files. That part is untested, and it may be quicker to run it over a big.matrix--
currently this just runs over a regular matrix before converting afterwards to
a big matrix.}

\item{axes}{How many axes from the PCA to retain and return.}
}
\value{
Nothing to the workspace. Saves PCA results to write.wd.
}
\description{
Combine a set of large eBird files and run a PCA over a specified set of columns.
}
\details{
It can be nearly impossible to run a standard base PCA over the
environmental variables in a set of eBird files. This function uses functions from
the data.table, bigmemory and bigpca packages to run a PCA. For small files it is
slower than a base R PCA, but it theoretically can handle very large inputs that
would normally crash R. I have a version of this file that runs in parallel, but not sure we need
it, and may swamp the RAM. Assess whether worth including parallel option at some
point. There is a major assumption built into the 'comparisons' list above, which
is that the files in read.wd begin with the species' names, with underscores
between the genus and species.
}

