% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEnv.R
\name{extractEnv}
\alias{extractEnv}
\title{Extract environmental data}
\usage{
extractEnv(ebird.data, env.wd, keep, longitude, latitude,
  cut.incomplete = TRUE)
}
\arguments{
\item{ebird.data}{Dataframe of eBird records.}

\item{env.wd}{Working directory where the environmental data is currently stored.
It would be very easy to make this able to download environmental data, set the
resolution you want in the data, and to use other forms
of environmental data besides bioclim, but those things are currently hardwired in.
It won't be much more work to make this more flexible when we're ready to take that
step.}

\item{keep}{Character vector of column names in the existing eBird records that you
would like to be passed along/kept in the cleaned files. Specify names(ebird.data),
(but substitute the actual name of your data frame, see example) to keep all columns.}

\item{longitude}{The name of the column in the eBird data files where the longitude
data is kept.}

\item{latitude}{The name of the column in the eBird data files where the latitude
data is kept.}

\item{cut.incomplete}{Default is TRUE. If a point is missing environmental data,
will remove that point before saving out the file.}
}
\value{
Saves data files with associated environmental data to the write.wd, returns
a data frame summarizing sample sizes to the workspace.
}
\description{
Extract and associate environmental data from GIS layers with eBird records.
}
\details{
There is a lot of room to make this function more flexible. Currently it
assumes you have bioclim data in a folder already on your computer at a resolution of
2.5 minutes. There is no reason the function cannot be generalized to download bioclim
data at whatever resolution you prefer, and there is no reason it cannot be generalized
to extract data from other GIS layers.
}
