% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypervolumeOverlaps.R
\name{hypervolumeOverlaps}
\alias{hypervolumeOverlaps}
\title{Calculate overlaps between all pairs of a hypervolume list}
\usage{
hypervolumeOverlaps(hv.list, what, cores)
}
\arguments{
\item{hv.list}{Object of class 'HypervolumeList'.}

\item{what}{Options are "jaccard" or "sorensen".}

\item{cores}{The number of cores to use for parallel processing.}
}
\value{
Matrix with rownames and colnames corresponding to the input hv.list, and
the elements filled in according to the specified 'what'.
}
\description{
Find all pairwise intersections, unions, etc., between elements in a hypervolume list.
}
\details{
Makes the assumption that the hv.list was loaded in with loadHypervolumes,
and that the Name elements in the list are of the form 'Genus_species.RDS'. Function
only calculates the set of {i,j}, not {j,i}. Because of the way
hypervolume_set works, these values can change slightly, so some precision is
lost here. Prints progress to screen.
}
\references{
Blonder, B., C. Lamanna, C. Violle, and B. J. Enquist. 2014.
The n-dimensional hypervolume. Global Ecology and Biogeography 23:595-609.
}
