% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHypervolumes.R
\name{loadHypervolumes}
\alias{loadHypervolumes}
\title{Load a set of hypervolumes}
\usage{
loadHypervolumes(read.wd, to.load)
}
\arguments{
\item{read.wd}{Working directory where the hypervolumes are stored.}

\item{to.load}{If left blank, will load all files in the read.wd. Otherwise, can
be used to only load a subset of species in the folder. Provide as a character
vector of the format 'Genus_species'.}
}
\value{
A hypervolume list with elements named according to the file names
in read.wd.
}
\description{
Load a set of hypervolumes into a single hypervolume list for plotting, etc.
}
\details{
The hypervolume package comes with a number of functions that can be used
over hypervolume list objects. The default plotting option, for instance, can be
rather useful.
}
