% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spThinner.R
\name{spThinner}
\alias{spThinner}
\title{Spatially thin a data frame of eBird records}
\usage{
spThinner(ebird.data, xxx, yyy, latitude, longitude, xlim = c(NA, NA),
  ylim = c(NA, NA), nx, ny, jitter = FALSE, size, replace = FALSE)
}
\arguments{
\item{ebird.data}{Dataframe of eBird records.}

\item{xxx}{Either E-W coordinates, or the word "extent". If extent is desired,
make sure to specify for both xxx & yyy. Function will not handle mixture gracefully.}

\item{yyy}{Either N-S coordinates, or the word "extent". If extent is desired,
make sure to specify for both xxx & yyy. Function will not handle mixture gracefully.}

\item{latitude}{The name of the latitude column in the files.}

\item{longitude}{The name of the longitude column in the files.}

\item{xlim}{Together xlim and ylim define a bounding box within which lookup occurs.
I.e. all (xxx, yyy) pairs outside of this box are ignored.}

\item{ylim}{Together xlim and ylim define a bounding box within which lookup occurs.
I.e. all (xxx, yyy) pairs outside of this box are ignored.}

\item{nx}{Number of grid cells in x direction.}

\item{ny}{Number of grid cells in y direction.}

\item{jitter}{Defaults to FALSE. Set to TRUE to randomize grid location.}

\item{size}{Maximum number of points per cell to sample}

\item{replace}{Defaults to FALSE. Set to TRUE to sample with replacement.}
}
\value{
A spatially thinned data frame.
}
\description{
Single data frame spatial thinning within an R session.
}
\details{
This applies sample.grid.cell() to single eBird data frame.
}
\examples{
data(ex)
plot(ex$LONGITUDE~ex$LATITUDE, col="blue", pch=20, cex=2)
thinned <- spThinner(ebird.data=ex, xxx="extent", yyy="extent",
  latitude="LATITUDE", longitude="LONGITUDE", nx=2, ny=2, size=3)
points(y=thinned$LONGITUDE, x=thinned$LATITUDE, col="red", pch=20, cex=1)
#the blue points are the original data, the red points are the accepted points
#after spatial thinning.
}
\references{
Team eBird.
}
