% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueFilter.R
\name{uniqueFilter}
\alias{uniqueFilter}
\title{Filter out group checklists.}
\usage{
uniqueFilter(dat, group.id, sub.id)
}
\arguments{
\item{dat}{The original eBird data.frame to filter.}

\item{group.id}{The name of the group ID column, usually "GROUP.ID".}

\item{sub.id}{The name of the sub ID column, or otherwise. Experimental,
have not really tested, but in theory can be used to sort dat so that, for
instance, only the first input of a set of group checklists is kept.}
}
\value{
A filtered data.frame.
}
\description{
Function to remove all but one of a set of eBird group checklists.
}
\details{
This function is called internally by cleaner(), but is also
exported here, as it may prove generally useful to some. Note that the results
are saved into the active R session, as opposed to written to file like the
cleaner() function.
}
\examples{
#load the example data file and delete group checklist duplicate entries
data(ex)
dim(ex)
ex2 <- uniqueFilter(ex, "GROUP_ID")
dim(ex2)
ex3 <- uniqueFilter(ex, "GROUP_ID", "SUB_ID")
dim(ex3)
setdiff(ex2$SUB_ID, ex3$SUB_ID)
}
\references{
Team eBird.
}

