% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEnv.R
\name{extractEnv}
\alias{extractEnv}
\title{Extract environmental data}
\usage{
extractEnv(env.wd, ebird.wd, write.wd, keep, longitude, latitude,
  cut.incomplete = TRUE, cores)
}
\arguments{
\item{env.wd}{Working directory where the environmental data is currently stored.
It would be very easy to make this able to download environmental data, set the
resolution you want in the data, and to use other forms
of environmental data besides bioclim, but those things are currently hardwired in.
It won't be much more work to make this more flexible when we're ready to take that
step.}

\item{ebird.wd}{Working directory where the eBird records are stored.}

\item{write.wd}{Working directory where the eBird records + environmental data will
be saved to.}

\item{keep}{Character vector of column names in the existing eBird records that you
would like to be passed along/kept in the cleaned files. Can be set to 'keep.all',
in which case all columns in the eBird data files will be kept.}

\item{longitude}{The name of the column in the eBird data files where the longitude
data is kept.}

\item{latitude}{The name of the column in the eBird data files where the latitude
data is kept.}

\item{cut.incomplete}{Default is TRUE. If a point is missing environmental data,
will remove that point before saving out the file.}

\item{cores}{The number of cores to use for parallel processing. This package currently
requires your computer to be able to handle at, the bare minimum, installation of the
various parallel processing packages. A forthcoming feature will allow you to specify
'seq' here, and would mean ebirdr could be run even if you can't install the parallel
processing software.}
}
\value{
Saves data files with associated environmental data to the write.wd, returns
a data frame summarizing sample sizes to the workspace.
}
\description{
Extract and associate environmental data from GIS layers with eBird records.
}
\details{
There is a lot of room to make this function more flexible. Currently it
assumes you have bioclim data in a folder already on your computer at a resolution of
2.5 minutes. There is no reason the function cannot be generalized to download bioclim
data at whatever resolution you prefer, and there is no reason it cannot be generalized
to extract data from other GIS layers.
}
