% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypervolumesToFile.R
\name{hypervolumesToFile}
\alias{hypervolumesToFile}
\title{Identify hypervolumes over multiple files}
\usage{
hypervolumesToFile(bandwidth.arg, r.points, species.col, trait.cols, read.wd,
  write.wd, cores)
}
\arguments{
\item{bandwidth.arg}{Options are either "silverman", "cross-validation", or a
fixed number, per standard hypervolume calculations. See Blonder et al. 2014 for more
details, including his FAQ page.}

\item{r.points}{The repsperpoint argument from hypervolume. If missing, will be set to
NULL and will default to hypervolume's default, 500 * the number of dimensions.}

\item{species.col}{The name of the column identifying the species to which the traits
belong. Not needed per se, but the function currently takes whatever column the user
provides here and binds the traits.col to it, ignoring this first column for the
hypervolume calculations. In other words, provide the name of some column here.}

\item{trait.cols}{A character vector detailing which columns from the eBird files you
would like to calculate a hypervolume for.}

\item{read.wd}{The path to the directory where the eBird records are housed.}

\item{write.wd}{The path to the directory where the hypervolumes will be saved.}

\item{cores}{The number of cores to use for parallel processing. This package currently
requires your computer to be able to handle at, the bare minimum, installation of the
various parallel processing packages. A forthcoming feature will allow you to specify
'seq' here, and would mean ebirdr could be run even if you can't install the parallel
processing software.}
}
\value{
Prints progress to the workspace, but saves nothing there. Results written
as RDS files to the write.wd.
}
\description{
Given a set of eBird files and specified traits, identify hypervolumes and save to disk.
}
\details{
Interesting and seemingly very useful method from Blonder et al. 2014, although
not entirely clear how best to set the bandwidth.arg parameters.
}
\references{
Blonder, B., C. Lamanna, C. Violle, and B. J. Enquist. 2014.
The n-dimensional hypervolume. Global Ecology and Biogeography 23:595-609.
}
