% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processModis.R
\name{processModis}
\alias{processModis}
\title{Process MODIS files from Google Earth Engine}
\usage{
processModis(species.col, read.wd, write.wd, aux.wd, keep, drop, cores)
}
\arguments{
\item{species.col}{Character vector specifying the column that contains the species
names. Will be used to split files.}

\item{read.wd}{Path to the read directory, where the files to be prepped are found.}

\item{write.wd}{Path to the write directory.}

\item{aux.wd}{Path to directory where the sample size summary file will be saved.}

\item{keep}{Character vector of columns to keep in the output files.}

\item{drop}{Should points that were not assigned complete MODIS values be dropped?
Default is TRUE.}

\item{cores}{How many cores to use for parallel processing.}
}
\value{
Nothing to workspace. Saves out two files. First to aux.wd is a
data frame summarizing per species sample sizes. Next are
species-level csv files with matched MODIS data. Spaces in the species names will
be removed.
}
\description{
Combine MODIS files and split into species-level files.
}
\details{
This function works by first binding all the files together in read.wd,
then dropping to all but the specified columns in keep, optionally dropping
incomplete records, then splitting into species-level files and saving out.
}
\references{
Team eBird.
}
